Updated the images to their original higher resolution and added color/filters to the overlay
The original images had to be reduced, maybe to fix the whole screen being taken over by images.  This is fixed by using overlay's not header.image.

Also got rid of the title on the home page so it doesn't overtake the overlay  image and repeat my name three times.